requirements.txt
matplotlib
pandas
pysine
pyalsaaudio
pyfiglet
